dic12021/requirements.txt
asgiref==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.6
django-leaflet==0.28.1
django-ranged-response==0.2.0
django-rest-knox==4.1.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
fontawesome-free==5.15.3
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.1.0
pycparser==2.20
python-gettext==4.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
